+++ business/package.json
@@ -28,10 +28,11 @@
   },
   "homepage": "https://github.com/ustaxcourt/ef-cms#readme",
   "dependencies": {
+    "aws-sdk": "^2.362.0",
     "axios": "^0.18.0",
     "joi": "^14.1.0",
-    "uuid": "^3.3.2",
-    "aws-sdk": "^2.362.0"
+    "sonar-scanner": "^3.1.0",
+    "uuid": "^3.3.2"
   },
   "devDependencies": {
     "chai": "^4.2.0",
